% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPack-package.R
\docType{package}
\name{AccessPack-package}
\alias{AccessPack}
\alias{AccessPack-package}
\title{AccessPack: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\references{
\url{https://github.com/soukhova/Spatial-Availability-Measures-Package}
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Antonio Paez, School of Earth, Environment and Society, McMaster University \email{paezha@mcmaster.ca}

Anastasia Soukhov, School of Earth, Environment and Society, McMaster University \email{soukhoa@mcmaster.ca}
}
\keyword{internal}
