% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPack-package.R
\docType{data}
\name{gtha_taz}
\alias{gtha_taz}
\title{gtha_taz}
\format{
A simple feature class (sf) polygon object containing 3764 rows and 5 variables; each row represents a unique TAZ with associated features.
\describe{
\item{GTA06}{Unique ID of the traffic analysis zone (TAZ).}
\item{AREA}{Area of TAZ in units of km^2.}
\item{workers}{The number of full-time workers within each TAZ.}
\item{jobs}{The number of jobs within each TAZd. }
\item{geometry}{The sfc polygon geometry (boundaries) of the TAZ.}
}
}
\source{
"2016 Transportation Tomorrow Survey" from \url{data management group} accessed November 14 2021.
}
\usage{
data(gtha_taz)
}
\description{
This object contains traffic analysis zones (TAZ) sourced from the 2016 Transportation Tomorrow Survey (TTS) for the entire Greater Toronto and Hamilton Area (GTHA).
}
\keyword{2016}
\keyword{GTHA}
\keyword{Hamilton}
\keyword{Jobs}
\keyword{TTS}
\keyword{Toronto}
\keyword{Workers}
