% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPack-package.R
\docType{data}
\name{hamilton_cma}
\alias{hamilton_cma}
\title{hamilton_cma}
\format{
A simple feature class (sfc) multipolygon object containing 1 row.
}
\source{
"City boundary" from \href{https://open.hamilton.ca/datasets/489f656072f44f34b07bf53efa0d8e4f_13/about}{Hamilton Open Data} accessed November 8 2021.
}
\usage{
data(hamilton_cma)
}
\description{
This object defines the spatial boundary of the City of Hamilton.
}
\keyword{CMA}
\keyword{Hamilton}
\keyword{boundary}
