% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPavailability.R
\name{p_allocation}
\alias{p_allocation}
\title{Multiply two numbers}
\usage{
p_allocation(x, o_id, d_id, pop, opp, f, alpha = 1)
}
\arguments{
\item{x}{A data frame with origin-destination information, including identifiers for origins, destinations, population, opportunities, and a pre-calculated impedance function}

\item{o_id}{A character string with the name of the column in data frame x that contains the unique origin identifiers}

\item{d_id}{A character string with the name of the column in data frame x that contains the unique destination identifiers}

\item{pop}{A character string with the name of the column in data frame x that contains the population}

\item{opp}{A character string with the name of the column in data frame x that contains the opportunities}

\item{f}{A character string with the name of the column in data frame x that contains the values of the impedance function}

\item{alpha}{A number string with the name of the column in data frame x that contains the values of the impedance function}
}
\value{
A vector with the number of opportunities available to o_id from d_id
}
\description{
Multiply two numbers
}
