% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPack-package.R
\docType{data}
\name{toronto_muni_bound}
\alias{toronto_muni_bound}
\title{toronto_muni_bound}
\format{
A simple feature class (sfc) multipolygon object containing 1 row.
}
\source{
"Regional Municipal boundary" from \href{https://open.toronto.ca/dataset/regional-municipal-boundary/}{Toronto Open Data} accessed November 20 2021.
}
\usage{
data(toronto_muni_bound)
}
\description{
This object defines the spatial boundary of the City of Toronto
}
\keyword{Municipal}
\keyword{Toronto}
\keyword{boundary}
