% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessPack-package.R
\docType{data}
\name{toy_trips}
\alias{toy_trips}
\title{toy_trips}
\format{
A dataframe containing 27 rows and 3 variables; each row represents the number of trips from "Population" to "Employment Center" with associated features.
\describe{
\item{Origin}{The unique ID for each "Population" from 1 to 9 }
\item{Destination}{The unique ID for each "Employment Center" from 1 to 3}
\item{Population}{The population corresponding to the "Population" ID.}
\item{trips}{The number of trips made from origin to destination.}
}
}
\usage{
data(toy_trips)
}
\description{
This object contains a hypothetical toy example of the number of trips for each Origin-Destination combination. Randomly generated.
}
\keyword{Destination}
\keyword{Jobs}
\keyword{Origin}
\keyword{Toy}
\keyword{Trips}
\keyword{Workers}
